Refactor summary API and update nvbench/summary.cuh docs.
The string used when constructing a summary is no longer a human
readable name, but rather a tag string (e.g. "nv/cold/time/gpu/mean").
These will make lookup easier and more stable going forward.

name vs. short_name no longer exists. Now there is just "name", which
is used for column headings. The "description" string may still be
used for detailed information.

Updated the json tests and compare script to reflect these changes.

+++ nvbench/type_axis.cxx
@@ -54,7 +54,8 @@ bool type_axis::get_is_active(std::size_t idx) const
 
 std::size_t type_axis::get_active_count() const
 {
-  return std::count(m_mask.cbegin(), m_mask.cend(), true);
+  return static_cast<std::size_t>(
+    std::count(m_mask.cbegin(), m_mask.cend(), true));
 }
 
 std::size_t type_axis::get_type_index(const std::string &input_string) const
